codequestionbot/backend/pyproject.toml
[project]
name = "codequestionbot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "cryptography>=46.0.3",
  "flask>=2.3",
  "flask-cors>=4.0",
  "psycopg[binary]>=3.1",
  "pyjwt>=2.10.1",
  "PyYAML>=6.0",
  "requests>=2.31",
  "SQLAlchemy>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
